refactor: create chaincode following Hyperledger's
Previously, we just created the chaincodes and call it a day, since we were just scaffolding the
first version of the blockchainizer.
However, now that we're implementing the chaincodes with the blockchain,
we are beginning to refactor the edges of the generator, so it matches what we want to do.

Hyperledger Fabric expects a npm package structure (that is, a package.json, the dist and src folder, etc...). Previously, the generator just created the TypeScript class and that's it. Now, we create the full package structure, as expected