refactor: remove dind labeling
Previously, for running DIND, we had to mount the host's Docker socket inside the container, which meant thatall the resources created inside the Docker container would be exposed to the host.

However, it's possible to run docker in docker completely isolated from the host (in the sense that it's not needed to share any resources for it, since privileged it's still needed), so all the handling logic for removing an specific label and labeling all resources upon creation is no longer needed.